legacy nonce issue fix
